requirements.txt
absl-py==1.4.0
ai-edge-litert==1.1.1
annotated-types==0.7.0
anyio==4.8.0
apache-beam==2.62.0
appdirs==1.4.4
argcomplete==3.5.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_record==0.6.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attr==0.3.1
attrs==25.1.0
avro-python3==1.10.2
azure-core==1.32.0
azure-storage-blob==12.24.1
babel==2.17.0
beautifulsoup4==4.13.3
betterproto==2.0.0b6
black==25.1.0
bleach==5.0.1
boto==2.49.0
boto3==1.36.18
botocore==1.36.18
Brotli==1.1.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.2.1
colorama==0.4.6
comm==0.2.2
contextlib2==21.6.0
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
datamodel-code-generator==0.26.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.1.1
distro==1.9.0
Django==4.2.19
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==2.4.0
django-migration-linter==5.1.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
dm-tree==0.1.9
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.16
docutils==0.21.2
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
email_validator==2.2.0
etils==1.11.0
executing==2.2.0
expiringdict==1.2.2
Faker==36.1.0
fastavro==1.10.0
fasteners==0.19
fastjsonschema==2.21.1
ffmpeg==1.4
filelock==3.17.0
flatbuffers==25.1.24
fonttools==4.55.7
fqdn==1.5.1
fsspec==2025.2.0
gast==0.6.0
genson==1.3.0
gin-config==0.5.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-appengine-logging==1.5.0
google-cloud-audit-log==0.3.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.4
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0rc1
grpc-google-iam-v1==0.14.0
grpcio==1.65.0
grpcio-status==1.70.0
grpclib==0.4.7
h11==0.14.0
h2==4.2.0
h5py==3.12.1
hdfs==2.7.3
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
humansignal-drf-yasg==1.21.10.post1
hyperframe==6.1.0
id==1.5.0
idna==3.10
ijson==3.3.0
immutabledict==4.2.1
importlib_metadata==8.5.0
importlib_resources==6.5.2
in-toto-attestation==0.9.3
inflect==5.6.2
inflection==0.5.1
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isodate==0.7.2
isoduration==20.11.0
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsf==0.11.2
json5==0.10.0
jsonpickle==3.4.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.6.17
kagglehub==0.3.7
keras==3.8.0
keyring==25.6.0
kiwisolver==1.4.8
label-studio==1.15.0
label-studio-sdk==1.0.8
launchdarkly-server-sdk==8.2.1
libclang==18.1.1
lockfile==0.12.2
lvis==0.5.3
lxml==5.3.0
lxml_html_clean==0.4.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.1
ml-dtypes==0.4.1
model-signing==0.2.0
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
mutagen==1.47.0
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.20
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cuda-nvcc-cu12==12.5.82
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauth2client==4.1.3
object_detection @ file:///home/jneval92/project_folder/train_classifier/models/research
objsize==0.7.1
openai==1.61.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opentelemetry-api==1.30.0
opt_einsum==3.4.0
optree==0.14.0
ordered-set==4.0.2
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pafy==0.5.5
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
portalocker==3.1.1
prometheus_client==0.21.1
promise==2.3
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyboxen==1.3.0
pycocotools==2.0.8
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydot==1.4.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11
pyOpenSSL==25.0.0
pyparsing==2.4.7
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pyRFC3339==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.4
python-slugify==8.0.4
pytz==2022.7.1
PyYAML==6.0.2
pyzmq==26.2.1
readme_renderer==44.0
redis==5.0.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-mock==1.12.1
requests-toolbelt==1.0.0
rfc3161-client==0.1.2
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rfc8785==0.1.4
rich==13.9.4
rpds-py==0.22.3
rq==1.10.1
rsa==4.9
rstr==3.2.2
rules==3.4
s3transfer==0.11.2
sacrebleu==2.2.0
scikit-learn==1.6.1
scipy==1.15.1
SecretStorage==3.3.3
securesystemslib==1.2.0
semver==3.0.4
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.21.0
seqeval==1.2.2
setuptools==75.8.0
sigstore==3.6.1
sigstore-protobuf-specs==0.3.2
sigstore-rekor-types==0.0.18
simple-parsing==0.1.7
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-datasets==4.9.7
tensorflow-hub==0.16.1
tensorflow-io==0.37.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
tensorflow-model-optimization==0.8.0
tensorflow-text==2.18.1
tensorflow_object_detection_api==0.1.1
termcolor==2.5.0
terminado==0.18.1
text-unidecode==1.3
tf-models-official==2.18.0
tf-slim==1.1.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
tuf==5.1.0
twine==6.1.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.10.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.1.3
wheel==0.40.0
widgetsnbextension==4.0.13
wrapt==1.17.2
xmljson==0.2.1
yt-dlp==2025.2.19
zipp==3.21.0
zstandard==0.23.0
